requirements.txt
marker-pdf
google-generativeai